+++ click_extra/__init__.py
@@ -52,6 +52,7 @@ from .config import (
     ConfigOption,
     NoConfigOption,
     ValidateConfigOption,
+    flatten_config_keys,
     get_tool_config,
     normalize_config_keys,
 )
@@ -193,6 +194,7 @@ __all__ = [
     "edit",
     "extraBasicConfig",
     "file_path",
+    "flatten_config_keys",
     "format_filename",
     "get_app_dir",
     "get_binary_stream",